fix(chain): simplify transaction validity period check (#2237)
Leverage doomslug finality to simplify the check for transaction validity period. Since under normal operation, a block becomes doomslug final just after one block, for blocks before the doomslug final height, we know that there cannot be more than one block for each height and it makes checking for transaction validity period straightforward. This allows us to have fairly large transaction validity period without compromising performance. Since doomslug has liveness and we project in reality the liveness threshold to be on the order of seconds, this approach can be considered as having O(1) complexity in practice.

This PR also removes the check for transaction validity in validating chunk proofs, which is inappropriate to begin with. A challenge should not depend on the chain information that a node has locally, as that would lead to very unfortunate consequences: a malicious actor can now arbitrarily challenge a chunk that is already garbage collected by non-archival nodes and those nodes would consider the chunk invalid because they don't have the blocks that far back to begin with and therefore the chunk producer will be slashed. On the other hand, the archival nodes will consider this challenge malicious and the creator of the challenge will be slashed. This essentially causes the network to diverge since some nodes would consider the challenge valid while others consider it malicious, which is really the worst that can happen to the network. Granted, there are other issues with challenges but they shall be addressed in separate PRs.

Test Plan
---------
Make sure existing tests for checking transaction validity period pass.