refactor(core): Fixed potentially dangerous implementation of parse_account_id_from_access_key_key (#2409)
*Resolves #2408*

I realized that I missed the fact that there is indeed a separator `col::ACCESS_KEY`, so the storage key is, essentially, constructed as `col::ACCESS_KEY + account_id + col::ACCESS_KEY + serialized_public_key`! I missed the fact that it is a constant there, which can be used instead of the magical `2`!

## Test Plan

No tests are needed, the magical `2` was the correct const.