fix(script): prepare for testnet update (#2428)
Update the migration script to protocol version 5 to make sure that it works for testnet.

Test plan
---------
Use the same genesis as currently running testnet to spin up a local node. After it produces blocks for a while, stop it and dump its state using the state-viewer in `stable`. Then apply `5-preserve-height.py`, `6-state-stake.py`, and `7-account-registar.py` to obtain the migrated genesis file. Then spin up a node with binary compiled from beta branch and make sure that the node can start and produce blocks.
@ailisp please independently verify that the migration scripts work.