fix(sync): fix state sync infinite loop (#2470)
Currently there is a potential infinite loop in state sync: if the node that is syncing cares about the shard, it will not be able to process sync block immediately, which will likely cause the node to go back into state sync after it is done (unless the sync block is close to header head), and because sync block is already removed at this point, future state syncs will get stuck as well after they are done. This PR fixes the issue by relaxing the criterion of entering block sync to allow block sync even if the head is one epoch behind. This is harmless in most scenarios, as a node would first do state sync anyways, unless it is right in the range of (one epoch, two epochs) behind header head.
Fixes #2374.

Test plan
---------
`state_sync3.py` that tests the scenario described above.