fix: correctly count chunk produced (#2535)
The current way of counting expected number of chunks for a chunk producer is flawed. While I took into consideration that when a block is not produced, not chunk can be included in that block, I failed to realize that if block at height `h` is not produced, chunks at height `h` can be included in block `h+1`, if it is produced. Therefore, what we really need to count when processing a block is whether chunk producers for height `prev_block_height + 1` produced their chunks. Resolves #2534.

Test plan
---------
`test_expected_chunks_prev_block_not_produced` that tests when a chunk producer failed to produce a chunk due to the block at previous height not being produced, it does not get kicked out.