fix: do not try to query local receipts (#2558)
When we query transaction status we sometimes need to route the query of receipts to other shard, but we should not query local receipts as they don't need to be queried and they are not saved in storage so querying it will cause a DBNotFound error. Fixes #2539. Fixes #2551 

Test plan
---------
* `gibberish.py` passes.
* `test_smart_contract_self_call_testnet` passes.