[WIP] Patches for RuntimeStandalone (#2626)
Some things turn out to be more clear being utilized.

For example, after adding `produce_blocks(num)` it turned out that `produce_block` in a better name for `process_block` since it better reflects semantics of what it actually does. 

This branch accumulates  patches which are produced in the process of utilizing RuntimeStandalone.