feat(rpc): add rpc for light client proof of execution outcomes (#2737)
Add a rpc endpoint `EXPERIMENTAL_light_client_proof` to extract a light client proof of execution outcome given a transaction hash or receipt id and the light client head. The rpc returns the following struct
```rust
pub struct RpcLightClientExecutionProofResponse {
    /// Proof of execution outcome
    pub outcome_proof: ExecutionOutcomeWithIdView,
    /// Proof of shard execution outcome root
    pub outcome_root_proof: MerklePath,
    /// A light weight representation of block that contains the outcome root
    pub block_header_lite: LightClientBlockLiteView,
    /// Proof of the existence of the block in the block merkle tree,
    /// which consists of blocks up to the light client head
    pub block_proof: MerklePath,
}
```
The rpc is broken into two parts: getting the execution outcome proof and getting the block proof. The former is relatively straightforward since we already store the proofs locally when transactions and receipts are executed. The latter involves constructing a merkle proof from partial merkle trees, which has a complexity of O((logn)^2) since for each node on the path, it takes O(logn) time to reconstruct. 


Test plan
---------
* `test_block_merkle_proof` that tests merkle proofs for blocks are correctly constructed.
* `test_get_execution_outcome` that tests execution outcome proof is correctly extracted
* pytest `rpc_light_client_execution_outcome_proof.py` that tests the rpc works end to end.
Note: due to #2584, the outcome root contains the errors which will require a very large schema to borsh serialize in python and therefore I am not able to test cases where transactions fail in pytest.