Fix(test): log is not shown completely when node crashed in pytest (#2763)
The root cause is once the node is panicked and exit, it takes a while for it to flush log to stdout and stderr (the pid still stays for one or two second before write finish, and even after os.kill it's there, until it write log finishes the node.pid disappear. So fix is in cleanup, check pid to ensure it's indeed finish. Fixes #2761 

Test Plan
----------
a short repo: add a if head.height>10 { panic!("aaa"); }, and run block_production.py. Before this PR log is truncated in test0_finished/stderr, but after this fix it's complete