fix(rpc): fix EXPERIMENTAL_check_tx (#2814)
As reported by @marcinbodnar, `EXPERIMENTAL_check_tx` returns `InternalError` when it is invoked. This is caused by a mistake I made when addressing some comment on the original PR (forgot to consider the variant in `send_tx`). 

Test plan
---------
Add test for `EXPERIMENTAL_check_tx` in `rpc_tx_submission.py` and make sure that it returns the expected error.