feat(metrics): Expose metrics friendly for dashboard (#2804)
Expose useful metrics in a friendly way to have a useful dashboard for on-fire situations.
As part of this effort worked on the dev-ops to deploy the dashboard. https://github.com/nearprotocol/near-ops/pull/53

There is currently a node on betanet running with this changes applied on top of beta branch. 
http://34.94.189.12:3030/status

# Testplan

Check that the [dashboard](http://35.236.63.163:3000/d/node-inspection/node-inspection?orgId=1&refresh=5s) is working properly. 

DISCLAIMER: While the node is syncing some graphs are not being displayed properly (since some information is not recorded). 

Once every node run this code, we are going to be able to select and explore each node individually using the dropdown from the upper left corner. Node will be added automatically as they join to the network.

<img width="1641" alt="Screen Shot 2020-06-06 at 2 03 14 AM" src="https://user-images.githubusercontent.com/28708963/83937384-08a42a00-a79a-11ea-8599-9ab9b6774075.png">

**UPDATE** This is the graph after the node finished syncing. More metrics can be displayed in demand:

<img width="1644" alt="Screen Shot 2020-06-06 at 2 16 36 PM" src="https://user-images.githubusercontent.com/28708963/83951595-7aaf5a00-a800-11ea-982b-8586e594db77.png">