fix: disable epoch info gc for now (#2952)
Currently we would garbage epoch information right after the epoch switch, which causes many problems including rpc not able to fetch chunks/blocks in the past epoch, not able to verify blocks in the past epoch, etc. I think it is safer to disable it for now and maybe consider implementing it properly when we redo the syncing code.