fix(chunks): process partial encoded chunk properly handles missing blocks (#3033)
The signature check for partial encoded chunks looks up the epoch id using the referenced block hash, however if this block is not known by the node then the check will fail, causing the processing of the chunk to return an error. However, the desired behaviour is to signal that a block is missing so that this chunk can be re-tried in the future. This fixes the issue, properly handling the potential error from the signature check and returning `NeedBlock` appropriately.

Note: The majority of the code is for the testing, the fix itself is quite small (see the second commit)

Testing Plan:

* New unit test for this case