fix(chain): always use the right chunk extra when producing chunks (#3056)
Currently we always use latest chunk extra when producing chunks. However, this might cause problems because the latest chunk extra is determined by the current head, which might not be the block that we are producing chunk on top of. When this happens, it causes invalid chunks to be produced, as mentioned in #3055. This PR fixes it by always using the chunk extra from the block that the chunk is building on. Fixes #3055.

Test plan
---------
`test_validate_chunk_extra`