fix(chunks): refactor `receipts_recipient_filter` to avoid excessive calls to `account_id_to_shard_id` (#3199)
This calling `account_id_to_shard_id` many times was causing a performance issue in `ShardsManager::distribute_encoded_chunk` because `receipts_recipient_filter` is called inside a loop, and `account_id_to_shard_id` has a non-trivial computation cost since it computes a hash.

This change is part of an ongoing effort to improve the performance of the node under load. This change does not "solve" the current performance issues, but it does give a small improvement.