fix: skip signature verification in produce_chunk (#3224)
When we produce a chunk, we have already verified the signature of transactions in the mempool and we don't need to check the same signature again. Even if the key is removed, we can skip the signature check and the access key check will correctly fail in this case. Also removes validation of transaction without state when we receive them. It doesn't make sense to verify signature twice there.

Test plan
---------
Existing tests