fix(mocknet): use the chain itself to measure bps and tps (#3248)
Previously blocks per second (bps) and transactions per second (tps) were computed based on metrics exposed by the node. However, these metrics may not always be an accurate measure (e.g. transactions will be double-counted in the case of a fork of the chain, and the bps metrics is an instantaneous snapshot rather than an average over a particular period of interest). Here, we change things so that bps and tps are measured based on the actual blockchain that is created by querying an archival node.

The `reset_data` function is also added to the mocknet library since it is useful when comparing different versions of the node or different configurations of the node (each run can start from a clean state to avoid possible interactions with existing state).