feat(indexer): Explicitly add local receipts to StreamerMessage (#3311)
As we found out some `AddKey` actions for transactions where signer is receiver (sir) the receipt created as "local" and is not included into `ChunkView`. But we need those receipts for indexing purposes altogether with `ExecutionOutcome` for such receipts. @bowenwang1996 addressed the issue https://github.com/nearprotocol/nearcore/issues/3305 to help us collect `ExecutionOutcomes` for specific block more easily (corresponding PR #3312) 

So in this update:

* `StreamerMessage` changes:
  * We add `local_receipts: Vec<near_primitives::views::ReceiptView>` to include local receipts
  * Rename `outcomes` field to `receipt_execution_outcomes`
  * `receipt_execution_outcomes` field type changed to be `HashMap<CryptoHash, near_primitives::views::ExecutionOutcomeWithIdView>` and includes **only Execution Outcomes for receipts**
  * `chunks` field type changes to be `Vec<IndexerChunkView>` where `transactions` field changes the type to `IndexerTransactionWithOutcome` (it is a wrapper around `near_primitives::views::SignedTransactionView`, it contains `transaction` itself and `outcome` for that transaction)