fix: always have partial encoded chunk when chunk exists (#3344)
Since #3036, nodes will persist all the partial encoded chunk parts for the shard they track. However, on mainnet there are quite a few blocks produced before the PR landed and therefore we need to do a migration for archival nodes that contain all the historical data.

Test plan
----------
Tested the migration on a mainnet archival node.