Fix(metrics): count blocks on block production and chunks on chunk production (#3943)
I noticed the `BLOCK_PRODUCED_TOTAL` was being incremented in `produce_chunk`. This is fine in the single shard setting where blocks and chunks are 1:1, but in the sharded setting this is incorrect. In this PR we increment `BLOCK_PRODUCED_TOTAL` when a block is produced and add a new metric `CHUNK_PRODUCED_TOTAL` to increment when a chunk is produced.